Shortcut to move yakuake to a next screen
With multiple screen setups sometimes you want to move yakuake to
another screen. Right now it's only possible from a menu which
makes it inconvenient.

This implements a shortcut that moves the yakuake window to a next
screen.
For example if you have 3 screens and start at screen 1:
* First invocation moves it to screen 2
* Second invocation moves it to screen 3
* Third invocation moves it to screen 1

If "At mouse location" is enabled, then the actual screen is used and
the window location is set to the next one in order. That means that to
select "At mouse location" again you have do it from the menu.